Fixed bugs where 0-valued indices were being stored in the vector and arrays
M    util/SparseHashArray.java

- Fixed bug in .set() where null-valued indices were being put in the backing
  map

M    vector/SparseHashDoubleVector.java
M    vector/SparseHashIntegerVector.java

- Fixed set calls to avoid double autoboxing

M    vector/SparseHashVector.java

- Updated to set all 0-valued indices as null in the backing array, which
  ensures the array stays sparse.

+++ trunk/sspace/src/edu/ucla/sspace/vector/SparseHashIntegerVector.java
@@ -95,7 +95,7 @@ public class SparseHashIntegerVector extends SparseHashVector<Integer>
      * {@inheritDoc}
      */
     public void set(int index, Number value) {
-        vector.set(index, value.intValue());
+        vector.set(index, value);
     }
 
     /**
